[test] Add separate testing files for `__int__`, `__float__`, `__str__` (#34)
This pull request includes several changes to the `DeciMojo` library,
focusing on code cleanup, new test additions, and minor optimizations.
The most important changes include the addition of a new test case, a
fix in the `Decimal` struct, and the removal of unnecessary comment
lines.

### New test additions:
*
[`mojoproject.toml`](diffhunk://#diff-1b0ef62120bccf4c05a76e60e13fe686f33eb92d5897b3f05e0c3f8d737fc5c0R43):
Added a new test case `test_to_float` to the test suite.

### Test removal:
*
[`tests/test_conversions.mojo`](diffhunk://#diff-d43bab99eda0af629ae4f094121805ed00644a6fa4ff98de5d068111ff7e7738L1-L129):
Removed the entire file containing tests for Decimal conversion methods
and split it into `test_to_string` and `test_to_int`.

### Code optimizations:
*
[`src/decimojo/decimal.mojo`](diffhunk://#diff-c2178e6e0bbbbb0d3379f1513412224309d1232c444654e77abb5ce36b3a854fL970-R968):
Fixed the `to_float` method in the `Decimal` struct to correctly use
`Float64(10)` for division, ensuring proper floating-point arithmetic.

### Documentation updates:
*
[`docs/todo.md`](diffhunk://#diff-d84697334acf64fe5d67dfd34f6370dca462f602aa3a720711eae12bf8dd153dL5-R5):
Marked the optimization task for the `exp()` function as completed.

+++ src/decimojo/decimal.mojo
@@ -1,5 +1,4 @@
 # ===----------------------------------------------------------------------=== #
-#
 # DeciMojo: A fixed-point decimal arithmetic library in Mojo
 # https://github.com/forFudan/DeciMojo
 #
@@ -16,7 +15,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
 # ===----------------------------------------------------------------------=== #
 #
 # Implements basic object methods for the Decimal type
@@ -967,7 +965,7 @@ struct Decimal(
             The floating-point representation of this Decimal.
         """
 
-        var result = Float64(self.coefficient()) / (10 ** self.scale())
+        var result = Float64(self.coefficient()) / (Float64(10) ** self.scale())
         result = -result if self.is_negative() else result
 
         return result
